+++ public/assets/js/adm_analisis_estructural_de_armaduras.js
@@ -1,11 +1,3 @@
-import { Grid } from "./cad/grid.js";
-import { History } from "./cad/history.js";
-import { Shape, Marker } from "./cad/shapes.js";
-import { pointDistance, distanceToSegment, formatCoordinate } from "./cad/utils.js";
-import { makeCreateDeleteColumn } from "./tabulator_base/table.js";
-import { createSpreeadSheetTable } from "./tabulator_base/table_factory.js";
-import { matlabColorScale } from "./matlab/color_scale.js";
-
 function getMousePos(canvas, evt) {
   const rect = canvas.getBoundingClientRect();
   const style = getComputedStyle(canvas);
@@ -27,326 +19,27 @@ function getMousePos(canvas, evt) {
 }
 
 document.addEventListener("DOMContentLoaded", () => {
-  const datosGeneralesModel = (id) => {
-    return {
-      id: id,
-      data: [
-        {
-          id: 1,
-          column: 27,
-          x: 0,
-          y: 0,
-          pd1: 5.1319,
-          pd2: -0.0348,
-          pd3: 0.0803,
-          pl1: 0.0912,
-          pl2: -0.0009,
-          pl3: 0.0001,
-          sismo1: 6.0318,
-          sismo2: 3.0391,
-          sismo3: 0.1554,
-        },
-        {
-          id: 2,
-          column: 1,
-          x: 0,
-          y: 14.13,
-          pd1: 10.2057,
-          pd2: 0.1979,
-          pd3: 0.1814,
-          pl1: 1.5233,
-          pl2: 0.1312,
-          pl3: 0.0191,
-          sismo1: 2.5603,
-          sismo2: 3.2251,
-          sismo3: 0.8937,
-        },
-        {
-          id: 3,
-          column: 3,
-          x: 4.8,
-          y: 14.13,
-          pd1: 9.8239,
-          pd2: 0.2463,
-          pd3: 0.0002,
-          pl1: 1.4297,
-          pl2: 0.1277,
-          pl3: 0.0185,
-          sismo1: 2.5856,
-          sismo2: 2.7967,
-          sismo3: 0.8873,
-        },
-        {
-          id: 4,
-          column: 5,
-          x: 0,
-          y: 10.87,
-          pd1: 13.3251,
-          pd2: -0.0238,
-          pd3: 0.1576,
-          pl1: 2.5238,
-          pl2: -0.0266,
-          pl3: 0.0146,
-          sismo1: 1.8216,
-          sismo2: 3.6588,
-          sismo3: 0.4599,
-        },
-        {
-          id: 5,
-          column: 7,
-          x: 4.8,
-          y: 10.87,
-          pd1: 18.7931,
-          pd2: 0.1901,
-          pd3: 0.0595,
-          pl1: 3.0877,
-          pl2: 0.0709,
-          pl3: 0.0257,
-          sismo1: 32.1112,
-          sismo2: 4.5443,
-          sismo3: 2.4735,
-        },
-        {
-          id: 6,
-          column: 12,
-          x: 0,
-          y: 7.23,
-          pd1: 14.7282,
-          pd2: -0.1881,
-          pd3: 0.1981,
-          pl1: 1.6038,
-          pl2: -0.1404,
-          pl3: 0.0089,
-          sismo1: 1.9516,
-          sismo2: 3.593,
-          sismo3: 0.419,
-        },
-        {
-          id: 7,
-          column: 13,
-          x: 4.8,
-          y: 7.23,
-          pd1: 14.7224,
-          pd2: -0.3165,
-          pd3: -0.1065,
-          pl1: 1.8861,
-          pl2: -0.2279,
-          pl3: 0.0054,
-          sismo1: 0.4018,
-          sismo2: 2.9011,
-          sismo3: 0.4681,
-        },
-        {
-          id: 8,
-          column: 15,
-          x: 6.72,
-          y: 7.23,
-          pd1: 9.9749,
-          pd2: 0.0000408,
-          pd3: -0.0061,
-          pl1: 0.807,
-          pl2: -0.0716,
-          pl3: 0.0027,
-          sismo1: 1.4632,
-          sismo2: 3.4036,
-          sismo3: 0.4395,
-        },
-        {
-          id: 9,
-          column: 17,
-          x: 0,
-          y: 2.73,
-          pd1: 9.4622,
-          pd2: -0.0227,
-          pd3: 0.0391,
-          pl1: -0.1035,
-          pl2: 0.0068,
-          pl3: 0.0023,
-          sismo1: 3.794,
-          sismo2: 3.6411,
-          sismo3: 0.2977,
-        },
-        {
-          id: 10,
-          column: 19,
-          x: 4.8,
-          y: 2.73,
-          pd1: 10.763,
-          pd2: -0.2191,
-          pd3: 0.0164,
-          pl1: 0.0193,
-          pl2: -0.0092,
-          pl3: 0.0033,
-          sismo1: 2.1638,
-          sismo2: 1.7708,
-          sismo3: 0.2303,
-        },
-        {
-          id: 11,
-          column: 21,
-          x: 6.72,
-          y: 2.73,
-          pd1: 6.3618,
-          pd2: 0.0513,
-          pd3: 0.0301,
-          pl1: -0.0489,
-          pl2: 0.0162,
-          pl3: 0.0047,
-          sismo1: 3.8355,
-          sismo2: 3.4022,
-          sismo3: 0.2552,
-        },
-        {
-          id: 12,
-          column: 23,
-          x: 4.8,
-          y: 0,
-          pd1: 8.2016,
-          pd2: -0.0442,
-          pd3: 0.0308,
-          pl1: 0.0553,
-          pl2: 0.0009,
-          pl3: 0.0035,
-          sismo1: 10.7601,
-          sismo2: 3.3956,
-          sismo3: 1.0017,
-        },
-        {
-          id: 13,
-          column: 25,
-          x: 6.72,
-          y: 0,
-          pd1: 5.4082,
-          pd2: 0.0524,
-          pd3: 0.0285,
-          pl1: -0.0155,
-          pl2: 0.0127,
-          pl3: 0.0034,
-          sismo1: 16.1751,
-          sismo2: 3.8987,
-          sismo3: 0.9936,
-        },
-        {
-          id: 14,
-          column: 166,
-          x: 6.72,
-          y: 10.87,
-          pd1: 10.6504,
-          pd2: 0.2522,
-          pd3: 0.05,
-          pl1: 0.9954,
-          pl2: 0.1076,
-          pl3: 0.0182,
-          sismo1: 32.9814,
-          sismo2: 4.4079,
-          sismo3: 2.3816,
-        },
-      ],
-      config: {
-        /* layout: "fitDataTable", */
-        height: 480,
-        columns: [
-          makeCreateDeleteColumn(id),
-          {
-            title: "Columna",
-            field: "column",
-            editor: "number",
-          },
-          {
-            title: "X",
-            field: "x",
-            editor: "number",
-          },
-          {
-            title: "Y",
-            field: "y",
-            editor: "number",
-          },
-          {
-            title: "PD",
-            columns: [
-              {
-                title: "",
-                field: "pd1",
-                editor: "number",
-              },
-              {
-                title: "",
-                field: "pd2",
-                editor: "number",
-              },
-              {
-                title: "",
-                field: "pd3",
-                editor: "number",
-              },
-            ],
-          },
-          {
-            title: "PL",
-            columns: [
-              {
-                title: "",
-                field: "pl1",
-                editor: "number",
-              },
-              {
-                title: "",
-                field: "pl2",
-                editor: "number",
-              },
-              {
-                title: "",
-                field: "pl3",
-                editor: "number",
-              },
-            ],
-          },
-          {
-            title: "SISMO",
-            columns: [
-              {
-                title: "",
-                field: "sismo1",
-                editor: "number",
-              },
-              {
-                title: "",
-                field: "sismo2",
-                editor: "number",
-              },
-              {
-                title: "",
-                field: "sismo3",
-                editor: "number",
-              },
-            ],
-          },
-        ],
-      },
-    };
-  };
 
-  const datosGenerales = createSpreeadSheetTable(datosGeneralesModel("#datosGenerales"));
+
+  
+  var VERSION = "3.2";
 
   // Init GUI Components
-  const canvas = document.querySelector("#plot canvas");
-
-  const form = document.querySelector("#plot form");
-  const text = document.querySelector("#plot textarea");
-  const ctx = canvas.getContext("2d");
+  var canvas = document.querySelector("#plot canvas");
+  var div = document.querySelector("#plot div");
+  var form = document.querySelector("#plot form");
+  var text = document.querySelector("#plot textarea");
+  var ctx = canvas.getContext("2d");
   const editor = document.getElementById("editor");
   const input = document.createElement("input");
   input.type = "number";
   input.style.color = "black";
   input.style.position = "absolute";
-  input.style.width = "100px";
   input.style.top = 0;
   input.style.left = 0;
   input.style.transform = "translate(-50%,-50%)";
-
   // Global vars
-  const Tools = {
+  var Tools = {
     MOVE: 0,
     LINE: 1,
     ADD: 2,
@@ -354,21 +47,9 @@ document.addEventListener("DOMContentLoaded", () => {
     ORIGIN: 4,
     VISIBILITY: 5,
     COLORIZE: 6,
-    SELECT: 7,
-    NONE: 8,
-  };
-  const Styles = {
-    ARRAY: 0,
-    ONE_ARRAY: 1,
-    NORMALIZED: 2,
-    ONE_NORMALIZED: 3,
   };
   const shapes = [];
-  let markers = [];
-
   var shape,
-    isDragging = false,
-    dragStart,
     grid,
     history,
     currentTool = Tools.LINE,
@@ -383,17 +64,23 @@ document.addEventListener("DOMContentLoaded", () => {
       };
       const distance = parseFloat(input.value);
       const newPoint = { x: last_point.x + unitVec.x * distance, y: last_point.y + unitVec.y * distance };
-      const isDone = shape.addPointToEnd(grid.worldToScreen(newPoint), grid);
+      const isDone = shape.addPointToEnd(newPoint);
       if (isDone) {
-        shape.calcularPropiedades();
         shapes.push(shape);
         editor.removeChild(input);
         shape = new Shape(true);
       }
     }
   });
+
+  var HANDLE_RELATIVE_RADIUS = 0.17; // Vertex handle radius relative to grid spacing
   var handleIsSelected = false;
   var selectedHandleIndex = 0;
+
+  var PEN_COMMAND = 9000;
+  var PEN_UP = 7000;
+  var PEN_COLOR0 = 8000;
+
   var COLORS = [
     "#000000", // 00 (Penup)
     "#2020FF", // 01 BLUE
@@ -413,33 +100,59 @@ document.addEventListener("DOMContentLoaded", () => {
     "#1E90FF", // 15 DODGERBLUE
     "#ADD8E6", // 16 LIGHTBLUE
   ];
+
   var current_selected_color = COLORS[1];
+
   // Functions
   function windowResize() {
     // Set actual size in memory (scaled to account for extra pixel density).
     const scale = window.devicePixelRatio; // Change to 1 on retina screens to see blurry canvas.
     canvas.width = parseFloat(getComputedStyle(canvas).width) * scale;
     canvas.height = parseFloat(getComputedStyle(canvas).height) * scale;
-    grid.set(parseInt(form.zoom.value), canvas);
+    grid.set(parseInt(form.zoom.value));
     redraw();
   }
-  function closestMarker(searchPoint) {
-    // Returns null if there are 0 points in the shape
-    var shortestDistance = 5;
-    for (let index = 0; index < markers.length; index++) {
-      const p = markers[index].point;
-      const distance = pointDistance(searchPoint, grid.worldToScreen(p));
-      if (distance <= shortestDistance) {
-        return markers[index];
-      }
-    }
+
+  //---------------------------------
+  // Distance to line segment.
+  // Adapted from http://stackoverflow.com/questions/849211/shortest-distance-between-a-point-and-a-line-segment
+  function sqr(x) {
+    return x * x;
   }
+  function pointDistanceSquared(v, w) {
+    return sqr(v.x - w.x) + sqr(v.y - w.y);
+  }
+  function pointDistance(v, w) {
+    return Math.sqrt(sqr(v.x - w.x) + sqr(v.y - w.y));
+  }
+  function distanceToSegmentSquared(p, v, w) {
+    var l2 = pointDistanceSquared(v, w);
+    if (l2 === 0) return pointDistanceSquared(p, v);
+    var t = ((p.x - v.x) * (w.x - v.x) + (p.y - v.y) * (w.y - v.y)) / l2;
+    if (t < 0) return pointDistanceSquared(p, v);
+    if (t > 1) return pointDistanceSquared(p, w);
+    return pointDistanceSquared(p, { x: v.x + t * (w.x - v.x), y: v.y + t * (w.y - v.y) });
+  }
+  function distanceToSegment(p, v, w) {
+    return Math.sqrt(distanceToSegmentSquared(p, v, w));
+  }
+  //---------------------------------
+
+  function formatCoordinate(coordinate) {
+    // Return shortest possible string representation, up to 2 digits of precision
+    //   9.00  -> "9"
+    //   9.10  -> "9.1"
+    //   9.12  -> "9.12"
+    //   9.123 -> "9.12"
+    return Number(coordinate.toFixed(2)).toString();
+  }
+
   function closestPoint(searchPoint) {
     // Returns null if there are 0 points in the shape
-    var shortestDistance = 5;
+    var shortestDistance = 0.65;
     for (let index = 0; index < shapes.length; index++) {
       const collided = shapes[index].points.find((p, index, points) => {
-        const distance = pointDistance(searchPoint, grid.worldToScreen(p));
+        const distance = pointDistance(searchPoint, p);
         return distance <= shortestDistance;
       });
       if (collided) {
@@ -447,55 +160,82 @@ document.addEventListener("DOMContentLoaded", () => {
       }
     }
   }
+
   function closestLine(searchPoint) {
-    var shortestDistance = 10;
+    var shortestDistance = 0.2;
     return shapes.find((s) => {
-      for (let index = 0; index < s.points.length; index++) {
-        const lineLength = pointDistance(grid.worldToScreen(s.points[index % s.points.length]), grid.worldToScreen(s.points[(index + 1) % s.points.length]));
-        const d1 = pointDistance(grid.worldToScreen(s.points[index % s.points.length]), searchPoint);
-        const d2 = pointDistance(grid.worldToScreen(s.points[(index + 1) % s.points.length]), searchPoint);
+      for (let index = 0; index < s.points.length - 1; index++) {
+        const lineLength = pointDistance(s.points[index], s.points[index + 1]);
+        const d1 = pointDistance(s.points[index], searchPoint);
+        const d2 = pointDistance(s.points[index + 1], searchPoint);
+        if (d1 + d2 >= lineLength - shortestDistance && d1 + d2 <= lineLength + shortestDistance) {
+          return true;
+        }
+      }
+      if (s.points.length > 2) {
+        const lineLength = pointDistance(s.points[0], s.points[s.points.length - 1]);
+        const d1 = pointDistance(s.points[0], searchPoint);
+        const d2 = pointDistance(s.points[s.points.length - 1], searchPoint);
         if (d1 + d2 >= lineLength - shortestDistance && d1 + d2 <= lineLength + shortestDistance) {
           return true;
         }
       }
     });
   }
+
   function undo() {
     shape.points = history.undo();
     redraw();
   }
+
   function redo() {
     shape.points = history.redo();
     redraw();
   }
+
   function switchTool(newTool) {
     currentTool = newTool;
     redraw();
   }
+
   function drawFinishedShape(s) {
     var i, p, color, line_color;
     ctx.save();
     // Draw lines
     if (s.points.length >= 2) {
-      line_color = "white";
+      if (s.points[0].color) {
+        line_color = COLORS[s.points[0].color];
+      } else {
+        line_color = "white";
+      }
       ctx.strokeStyle = line_color;
-      ctx.lineWidth = 1;
+      ctx.lineWidth = 2;
       ctx.beginPath();
       for (i = 0; i < s.points.length; i++) {
-        p = grid.worldToScreen(s.points[i]);
+        p = grid.toPoint(s.points[i]);
         if (i === 0) {
           ctx.moveTo(p.x, p.y);
           continue;
         }
         ctx.lineTo(p.x, p.y);
         ctx.stroke();
-        ctx.strokeStyle = line_color;
-        ctx.setLineDash([]);
-        ctx.beginPath();
-        ctx.moveTo(p.x, p.y);
+        if (p.color) {
+          line_color = COLORS[p.color];
+        }
+        if (p.visible) {
+          ctx.strokeStyle = line_color;
+          ctx.setLineDash([]);
+          ctx.beginPath();
+          ctx.moveTo(p.x, p.y);
+        } else {
+          ctx.strokeStyle = "gray";
+          ctx.setLineDash([5, 10]);
+          ctx.beginPath();
+          ctx.moveTo(p.x, p.y);
+        }
       }
       if (s.points.length > 2) {
-        const begin = grid.worldToScreen(s.points[0]);
+        const begin = grid.toPoint(s.points[0]);
         ctx.lineTo(begin.x, begin.y);
         ctx.stroke();
       }
@@ -511,30 +251,30 @@ document.addEventListener("DOMContentLoaded", () => {
       } else if (i == 0) {
         color = "cyan";
       } else {
-        color = "red";
+        color = "green";
       }
-      p = grid.worldToScreen(s.points[i]);
+      p = grid.toPoint(s.points[i]);
       ctx.fillStyle = color;
       ctx.strokeStyle = color;
       ctx.beginPath();
-      ctx.arc(p.x, p.y, grid.size, 0, 2 * Math.PI);
+      ctx.arc(p.x, p.y, HANDLE_RELATIVE_RADIUS * grid.size, 0, 2 * Math.PI);
       ctx.fill();
       ctx.stroke();
     }
     ctx.restore();
   }
+
+  function drawUnfinishedShape(s) {}
+
   function redraw() {
     var x_pos = 10;
     var y_pos = 15;
     var y_step = 15;
-    grid.draw(ctx);
-    markers.forEach((m) => {
-      m.draw(grid, ctx);
-    });
+    grid.draw();
     shapes.forEach((s) => {
       drawFinishedShape(s);
     });
-    shape.draw(grid, ctx);
+    shape.draw(grid);
 
     ctx.save();
     ctx.font = "12px Helvetica";
@@ -548,20 +288,33 @@ document.addEventListener("DOMContentLoaded", () => {
     ctx.fillStyle = "white";
     ctx.fillText("(" + formatCoordinate(mousePos.x) + ", " + formatCoordinate(mousePos.y) + ")", x_pos, y_pos);
     y_pos += y_step;
+    const mPos = grid.toPoint(mousePos);
+    ctx.fillText("(" + formatCoordinate(mPos.x) + ", " + formatCoordinate(mPos.y) + ")", x_pos, y_pos);
+    y_pos += y_step;
+    const lp = shape.getLastPoint();
+    if (lp) {
+      ctx.fillText("(" + formatCoordinate(lp.x) + ", " + formatCoordinate(lp.y) + ")", x_pos, y_pos);
+      y_pos += y_step;
+      const midPoint = {
+        x: (mPos.x + lp.x) * 0.5,
+        y: (mPos.y + lp.y) * 0.5,
+      };
+      ctx.fillText("(" + formatCoordinate(midPoint.x) + ", " + formatCoordinate(midPoint.y) + ")", x_pos, y_pos);
+      y_pos += y_step;
+    }
 
     // Show current tool
     var modeText = "";
     switch (currentTool) {
       case Tools.MOVE:
         modeText = "Move";
-        /* selectedHandleIndex.drawTranslated(position); */
         break;
       case Tools.LINE:
         modeText = "Line";
         var last_point = shape.getLastPoint();
         if (last_point) {
-          var c = grid.worldToScreen(last_point);
-          const mouse = grid.worldToScreen(mousePos);
+          var c = grid.toPoint(last_point);
+          const mouse = grid.toPoint(mousePos);
           ctx.setLineDash([]);
           ctx.strokeStyle = "gray";
           ctx.beginPath();
@@ -588,279 +341,539 @@ document.addEventListener("DOMContentLoaded", () => {
     }
     ctx.fillText(modeText, x_pos, y_pos);
     ctx.restore();
-
-    document.getElementById("polygons").innerHTML = `
-         ${shapes.reduce((body, shape, index) => {
-           const propiedades = shape.propiedades();
-           return (
-             body +
-             `<tr class="bg-gray-100 dark:bg-gray-600">
-                  <td class="p-0" colspan="4">
-                      <div class="relative inline-block">
-                          <table class="inline-block text-gray-800 dark:text-white">
-                              <tbody id="polygons">
-                                  <tr
-                                      class="bg-white text-gray-900 dark:bg-gray-800 dark:text-white">
-                                      <th class="text-xl py-2 px-4 text-left" colspan="4">
-                                          Propiedades
-                                      </th>
-                                  </tr>
-                                  <tr
-                                      class="bg-gray-500 text-white dark:bg-gray-500 dark:text-white">
-                                      <th class="text-lg py-2 px-8" scope="col"
-                                          colspan="2">P</th>
-                                      <td class="py-2 px-4">${propiedades.P.toFixed(2)}</td>
-                                  </tr>
-                                  <tr
-                                      class="bg-gray-500 text-white dark:bg-gray-500 dark:text-white">
-                                      <th class="text-lg py-2 px-8" scope="col"
-                                          colspan="2">A</th>
-                                      <td class="py-2 px-4">${propiedades.A.toFixed(2)}</td>
-                                  </tr>
-                                  <tr
-                                      class="bg-gray-500 text-white dark:bg-gray-500 dark:text-white">
-                                      <th class="text-lg py-2 px-8" scope="col"
-                                          colspan="2">IX</th>
-                                      <td class="py-2 px-4">${propiedades.IX.toFixed(2)}</td>
-                                  </tr>
-                                  <tr
-                                      class="bg-gray-500 text-white dark:bg-gray-500 dark:text-white">
-                                      <th class="text-lg py-2 px-8" scope="col"
-                                          colspan="2">
-                                          IY</th>
-                                      <td class="py-2 px-4">${propiedades.IY.toFixed(2)}</td>
-                                  </tr>
-                                  <tr
-                                      class="bg-gray-500 text-white dark:bg-gray-500 dark:text-white">
-                                      <th class="text-lg py-2 px-8" scope="col"
-                                          colspan="2">
-                                          XC</th>
-                                      <td class="py-2 px-4">${propiedades.XC.toFixed(2)}</td>
-                                  </tr>
-                                  <tr
-                                      class="bg-gray-500 text-white dark:bg-gray-500 dark:text-white">
-                                      <th class="text-lg py-2 px-8" scope="col"
-                                          colspan="2">
-                                          YC</th>
-                                      <td class="py-2 px-4">${propiedades.YC.toFixed(2)}</td>
-                                  </tr>
-                                  <tr
-                                      class="bg-gray-500 text-white dark:bg-gray-500 dark:text-white">
-                                      <th class="text-lg py-2 px-8" scope="col"
-                                          colspan="2">
-                                          MX</th>
-                                      <td class="py-2 px-4">${propiedades.MX.toFixed(2)}</td>
-                                  </tr>
-                                  <tr
-                                      class="bg-gray-500 text-white dark:bg-gray-500 dark:text-white">
-                                      <th class="text-lg py-2 px-8" scope="col"
-                                          colspan="2">
-                                          MY</th>
-                                      <td class="py-2 px-4">${propiedades.MY.toFixed(2)}</td>
-                                  </tr>
-                                  <tr
-                                      class="bg-gray-500 text-white dark:bg-gray-500 dark:text-white">
-                                      <th class="text-lg py-2 px-8" scope="col"
-                                          colspan="2">
-                                          IXY</th>
-                                      <td class="py-2 px-4">${propiedades.IXY.toFixed(2)}</td>
-                                  </tr>
-                              </tbody>
-                          </table><table
-                              class="inline-block text-gray-800 dark:text-white absolute overflow-y-auto top-0 bottom-0">
-                              <tr class="bg-white text-gray-900 dark:bg-gray-800 dark:text-white">
-                                <th class="text-xl py-2 px-4 text-left" colspan="4">Poligono ${index + 1}
-                                </th>
-                              </tr>
-                              <tr class="bg-gray-500 text-white dark:bg-gray-500 dark:text-white">
-                                <th class="text-lg py-2 px-8" scope="col">X
-                                </th>
-                                <th class="text-lg py-2 px-4" scope="col">Y</th>
-                              </tr>
-                              ${shape.points.reduce((body, p) => {
-                                return (
-                                  body +
-                                  `
-                                  <tr class="bg-gray-100 dark:bg-gray-600">
-                                    <td class="py-2 px-4">${p.x.toFixed(2)}</td>
-                                    <td class="py-2 px-4">${p.y.toFixed(2)}</td>
-                                  </tr>`
-                                );
-                              }, "")}
-                          </table>
-                      </div>
-                  </td>
-              </tr>`
-           );
-         }, "")}
-          `;
   }
+
+  // DataObjs
+  var Shape = function (parseUrl) {
+    this.reset(parseUrl);
+  };
+  Shape.prototype = {
+    reset: function (parseUrl) {
+      this.points = [];
+
+      if (parseUrl) {
+        // Can supply initial point data in URL following "?".
+        // For example: wwww.mylocation.com/polygondraw.html?[0,0,1,1]
+        var pointsString = window.location.search.substring(1);
+        var color_index = null;
+        var pending_color = null;
+        if (pointsString !== "") {
+          var pointArray = JSON.parse(pointsString);
+          if (pointArray.length % 2 === 0) {
+            for (var i = 0, len = pointArray.length; i < len; i += 2) {
+              if (pointArray[i] == PEN_COMMAND) {
+                if (pointArray[i + 1] == PEN_UP) {
+                  this.points[this.points.length - 1].visible = false;
+                } else {
+                  color_index = pointArray[i + 1] - PEN_COLOR0;
+                  if (color_index >= 0 && color_index < COLORS.length) {
+                    if (i === 0) {
+                      // If a color is specified as the very fist directive,
+                      // then that color is applied to the next vertex (which
+                      // will be the fist vertex in the data set).  Otherwise,
+                      // color directives are applied to the previous vertex.
+                      pending_color = color_index;
+                    } else {
+                      this.points[this.points.length - 1].color = color_index;
+                    }
+                  }
+                }
+              } else {
+                this.points.push({
+                  x: pointArray[i],
+                  y: pointArray[i + 1],
+                  visible: true,
+                  color: pending_color ? pending_color : null,
+                });
+                pending_color = null;
+              }
+            }
+            // Start in MOVE mode if starting with an existing dataset
+            currentTool = Tools.MOVE;
+          }
+        }
+      }
+    },
+
+    addPointToEnd: function (position) {
+      /* const last_point = this.getLastPoint(); */
+      const begin = this.points[0];
+      if (this.points.length != 0 && pointDistance(begin, position) <= 0.65) {
+        // No noting if the add location is the same as the last point
+        return true;
+      }
+      this.addPointAfterIndex(this.points.length - 1, position);
+      return false;
+    },
+
+    addPointAfterIndex: function (index, position) {
+      shape.points.splice(index + 1, 0, { x: position.x, y: position.y, visible: true, color: null });
+      this._optimize();
+    },
+
+    getLastPoint: function () {
+      if (this.points.length === 0) {
+        return null;
+      }
+
+      this.last_point = this.points[this.points.length - 1];
+      return {
+        x: this.last_point.x,
+        y: this.last_point.y,
+      };
+    },
+
+    deletePoint: function (index) {
+      this.points.splice(index, 1);
+      this._optimize();
+    },
+
+    colorizePoint: function (index, color) {
+      this.points[index].color = color;
+      this._optimize();
+    },
+
+    _optimize: function () {
+      // Optimize the color assignments.  Color only needs to be specified once, on
+      // the first node that the color is drawn TO (not from)
+      var i, point, color;
+      color = null;
+      for (i = 0; i < this.points.length; i++) {
+        point = this.points[i];
+        if (color !== null && point.color == color) {
+          point.color = null;
+        }
+        if (point.color !== null) {
+          color = point.color;
+        }
+      }
+    },
+
+    draw: function (grid_) {
+      var i, p, color, line_color;
+
+      ctx.save();
+
+      // Draw lines
+      if (this.points.length >= 2) {
+        if (this.points[0].color) {
+          line_color = COLORS[this.points[0].color];
+        } else {
+          line_color = "white";
+        }
+        ctx.strokeStyle = line_color;
+        ctx.lineWidth = 2;
+        ctx.beginPath();
+        for (i = 0; i < this.points.length; i++) {
+          p = grid_.toPoint(this.points[i]);
+          if (i === 0) {
+            ctx.moveTo(p.x, p.y);
+            continue;
+          }
+          ctx.lineTo(p.x, p.y);
+          ctx.stroke();
+
+          if (p.color) {
+            line_color = COLORS[p.color];
+          }
+
+          if (p.visible) {
+            ctx.strokeStyle = line_color;
+            ctx.setLineDash([]);
+            ctx.beginPath();
+            ctx.moveTo(p.x, p.y);
+          } else {
+            ctx.strokeStyle = "gray";
+            ctx.setLineDash([5, 10]);
+            ctx.beginPath();
+            ctx.moveTo(p.x, p.y);
+          }
+        }
+        ctx.stroke();
+      }
+
+      // Draw vertex handles
+      for (i = 0; i < this.points.length; i++) {
+        if (handleIsSelected && i === selectedHandleIndex) {
+          color = "red";
+        } else if (i == this.points.length - 1) {
+          color = "blue";
+        } else if (i == 0) {
+          color = "cyan";
+        } else {
+          color = "green";
+        }
+        p = grid_.toPoint(this.points[i]);
+        ctx.fillStyle = color;
+        ctx.strokeStyle = color;
+        ctx.beginPath();
+        ctx.arc(p.x, p.y, HANDLE_RELATIVE_RADIUS * grid.size, 0, 2 * Math.PI);
+        ctx.fill();
+        ctx.stroke();
+      }
+      ctx.restore();
+    },
+  };
+
+  var History = function () {
+    this.clear();
+  };
+
+  History.prototype = {
+    clear: function () {
+      this.pointsHistory = [];
+      this.tail = -1;
+    },
+
+    commit: function (points) {
+      // Remove any history after the current tail
+      var splice_index = this.tail + 1;
+      if (splice_index <= this.pointsHistory.length - 1) {
+        this.pointsHistory.splice(splice_index, this.pointsHistory.length - splice_index);
+      }
+      // Add the new point set to the history
+      var clone = this.clone(points);
+      this.pointsHistory.push(this.clone(points));
+      this.tail = this.pointsHistory.length - 1;
+    },
+
+    undo: function () {
+      --this.tail;
+      if (this.tail <= -1) {
+        this.tail = -1;
+        return [];
+      } else {
+        return this.clone(this.pointsHistory[this.tail]);
+      }
+    },
+
+    redo: function () {
+      ++this.tail;
+      if (this.tail > this.pointsHistory.length - 1) {
+        this.tail = this.pointsHistory.length - 1;
+        console.log("Nothing to redo.");
+      }
+      if (this.tail === -1) {
+        return [];
+      } else {
+        return this.clone(this.pointsHistory[this.tail]);
+      }
+    },
+
+    clone: function (points) {
+      var pointsClone = [];
+      for (var i = 0, len = points.length; i < len; i++) {
+        pointsClone.push({
+          x: points[i].x,
+          y: points[i].y,
+          visible: points[i].visible,
+          color: points[i].color,
+        });
+      }
+      return pointsClone;
+    },
+  };
+
+  var Grid = function () {
+    this.canvas = document.createElement("canvas");
+    this.ctx = this.canvas.getContext("2d");
+  };
+  Grid.prototype = {
+    set: function (size) {
+      this.size = size;
+
+      var wt = Math.ceil(canvas.width / size);
+      var ht = Math.ceil(canvas.height / size);
+      this.width = (wt + (wt % 2)) * size + 1;
+      this.height = (ht + (ht % 2)) * size + 1;
+
+      this.x = 0;
+      this.y = 0;
+
+      /*       this.x = Math.round((canvas.width - this.width) / 2);
+      this.y = Math.round((canvas.height - this.height) / 2); */
+
+      this.origo = {
+        x: Math.round(this.width / (this.size * 2)),
+        y: Math.round(this.height / (this.size * 2)),
+      };
+
+      this.canvas.width = this.width;
+      this.canvas.height = this.height;
+
+      this.createGrid();
+      this.draw();
+    },
+
+    createGrid: function () {
+      var x = 0,
+        y = 0,
+        w = this.width - 1,
+        h = this.height - 1;
+
+      this.ctx.save();
+      this.ctx.strokeStyle = "#282828";
+      this.ctx.fillStyle = "#000";
+      this.ctx.fillRect(0, 0, this.width, this.height);
+
+      this.ctx.beginPath();
+      this.ctx.rect(x, y, w, h);
+      while (x < w) {
+        x += this.size;
+        this.ctx.moveTo(x, y);
+        this.ctx.lineTo(x, y + h);
+      }
+      x = 0;
+      while (y < h) {
+        y += this.size;
+        this.ctx.moveTo(x, y);
+        this.ctx.lineTo(x + w, y);
+      }
+      this.ctx.stroke();
+
+      this.ctx.beginPath();
+      this.ctx.fillStyle = "#383838";
+      this.ctx.arc(this.origo.x * this.size, this.origo.y * this.size, 5, 0, 2 * Math.PI);
+      this.ctx.fill();
+
+      this.ctx.restore();
+    },
+
+    draw: function () {
+      ctx.drawImage(this.canvas, this.x, this.y);
+    },
+
+    translate: function (x, y) {
+      if (snap_enabled) {
+        // Snap to grid
+        x = Math.round((x - this.x) / this.size);
+        y = Math.round((y - this.y) / this.size);
+      } else {
+        // Free form
+        x = (x - this.x) / this.size;
+        y = (y - this.y) / this.size;
+      }
+
+      return {
+        x: x - this.origo.x,
+        y: this.origo.y - y,
+      };
+    },
+
+    toPoint: function (p) {
+      return {
+        x: (p.x + this.origo.x) * this.size + this.x,
+        y: (this.origo.y - p.y) * this.size + this.y,
+        visible: p.visible,
+        color: p.color,
+      };
+    },
+  };
+
+  // Setup
   shape = new Shape(true);
   grid = new Grid();
   history = new History();
+
   windowResize();
   redraw();
+
   window.onresize = windowResize;
+
   form.zoom.oninput = function () {
     // Responds during slide Firefox, Safari, Chrome
     ctx.clearRect(0, 0, canvas.width, canvas.height);
-    grid.set(parseInt(this.value), canvas);
+    grid.set(parseInt(this.value));
     redraw();
   };
+
   form.zoom.onchange = function () {
     // IE10 support (IE10 does not support oninput)
     ctx.clearRect(0, 0, canvas.width, canvas.height);
-    grid.set(parseInt(this.value), canvas);
+    grid.set(parseInt(this.value));
     redraw();
   };
-  canvas.addEventListener(
-    "wheel",
-    (e) => {
-      e.preventDefault();
-      const { x, y } = getMousePos(canvas, e);
-      const prevMouse = grid.screenToWorld({ x: x, y: y });
-      if (e.deltaY < 0) {
-        grid.scaleX *= 1.1;
-        grid.scaleY *= 1.1;
-      } else {
-        grid.scaleX *= 0.9;
-        grid.scaleY *= 0.9;
-      }
-      const translatedMouse = grid.screenToWorld({ x: x, y: y });
-      grid.offestX += prevMouse.x - translatedMouse.x;
-      grid.offestY += prevMouse.y - translatedMouse.y;
-      redraw();
-    },
-    { passive: false }
-  );
+
   canvas.onmousedown = function (evt) {
-    evt.preventDefault();
     const { x, y } = getMousePos(canvas, evt);
+    if (0 > x || canvas.width < x || 0 > y || canvas.height < y) {
+      return;
+    }
+
+    var click = grid.translate(x, y);
+    var index, i, len;
     switch (currentTool) {
       case Tools.LINE:
-        const isDone = shape.addPointToEnd(grid.worldToScreen(mousePos), grid);
+        const isDone = shape.addPointToEnd(click);
         if (shape.points.length >= 1) {
           editor.appendChild(input);
         }
         if (isDone) {
-          shape.calcularPropiedades();
           shapes.push(shape);
           editor.removeChild(input);
+          /* document.getElementById("polygons").innerHTML += `
+          <tr class="bg-white text-gray-900 dark:bg-gray-800 dark:text-white">
+            <th class="text-xl py-2 px-4 text-left" colspan="4">Poligono ${shapes.length}
+            </th>
+          </tr>
+          <tr class="bg-gray-500 text-white dark:bg-gray-500 dark:text-white">
+            <th class="text-lg py-2 px-8" scope="col">X
+            </th>
+            <th class="text-lg py-2 px-4" scope="col">Y</th>
+          </tr>
+          ${shape.points.reduce((body, p) => {
+            return (
+              body +
+              `
+              <tr class="bg-gray-100 dark:bg-gray-600">
+                <td class="py-2 px-4">${p.x.toFixed(2)}</td>
+                <td class="py-2 px-4">${p.y.toFixed(2)}</td>
+              </tr>`
+            );
+          }, "")}
+          `; */
           shape = new Shape(true);
         }
         history.commit(shape.points);
         break;
+
       case Tools.MOVE:
-        if (evt.button == 1 || 1 == (evt.button & 2)) {
-          isDragging = true;
-          dragStart = { x: x, y: y };
-          return;
-        }
         if (handleIsSelected) {
           handleIsSelected = false;
-          let marker = closestMarker({ x: x, y: y });
-          selectedHandleIndex.calcularPropiedades();
-          const { XC: xc, YC: yc } = selectedHandleIndex.propiedades();
-          const dX = marker.point.x - xc;
-          const dY = marker.point.y - yc;
-          selectedHandleIndex.points.forEach((point) => {
-            point.x += dX;
-            point.y += dY;
-          });
           history.commit(shape.points);
         } else {
-          let index = closestPoint({ x: x, y: y });
+          index = closestPoint(click);
           if (index) {
             handleIsSelected = true;
             selectedHandleIndex = index;
-          } else if ((index = closestLine({ x: x, y: y }))) {
+          } else if ((index = closestLine(click))) {
             handleIsSelected = true;
             selectedHandleIndex = index;
           }
         }
         break;
+
+      case Tools.ADD:
+        len = shape.points.length;
+        var shortestDistance = 10000000000;
+        var matchIndex = 0;
+        if (len >= 2) {
+          // Find the nearest line segment to the click
+          for (i = 0; i < len - 1; i++) {
+            var distance = distanceToSegment(click, shape.points[i], shape.points[i + 1]);
+            if (distance < shortestDistance) {
+              shortestDistance = distance;
+              matchIndex = i;
+            }
+          }
+          // Add the new point
+          shape.addPointAfterIndex(matchIndex, click);
+          history.commit(shape.points);
+        }
+        break;
+
       case Tools.CUT:
-        const deleteShape = closestLine({ x: x, y: y });
-        const index = shapes.indexOf(deleteShape);
-        if (index !== -1) {
-          shapes.splice(index, 1);
+        index = closestPoint(click);
+        shape.deletePoint(index);
+        history.commit(shape.points);
+        break;
+
+      case Tools.ORIGIN:
+        // Move the origin to the click location
+        for (i = 0, len = shape.points.length; i < len; i++) {
+          shape.points[i].x -= click.x;
+          shape.points[i].y -= click.y;
+        }
+        history.commit(shape.points);
+        break;
+      case Tools.VISIBILITY:
+        // Toggle visibility of closest point
+        index = closestPoint(click);
+        if (index !== null) {
+          // First point may not be invisible
+          // Two consecutive points may not be invisible
+          // Last and next-to-last point may not be invisible
+          if (
+            index > 0 &&
+            shape.points[index - 1].visible &&
+            !(shape.points.length - 1 > index && !shape.points[index + 1].visible) &&
+            index < shape.points.length - 2
+          ) {
+            shape.points[index].visible = !shape.points[index].visible;
+          }
+        }
+        history.commit(shape.points);
+        break;
+      case Tools.COLORIZE:
+        // Toggle visibility of closest point
+        index = closestPoint(click);
+        if (index !== null) {
+          shape.colorizePoint(index, current_selected_color);
+          history.commit(shape.points);
         }
         break;
     }
 
     redraw();
   };
-  canvas.onmouseup = (evt) => {
-    if (evt.button == 1 || 1 == (evt.button & 2)) {
-      isDragging = false;
-    }
-  };
-  canvas.onmouseleave = () => {
-    isDragging = false;
-  };
+
   canvas.onmousemove = function (evt) {
     const { x, y } = getMousePos(canvas, evt);
-    mousePos = grid.screenToWorld({ x: x, y: y }, snap_enabled);
-    if (snap_enabled) {
-      mousePos.x = Math.floor((mousePos.x + 0.5) * grid.gridSpacing);
-      mousePos.y = Math.floor((mousePos.y + 0.5) * grid.gridSpacing);
+    if (0 > x || canvas.width < x || 0 > y || canvas.height < y) {
+      return;
     }
+    const prevMouse = mousePos;
+    mousePos = grid.translate(x, y);
+    redraw();
+
     if (currentTool === Tools.LINE && shape) {
       const last_point = shape.getLastPoint();
       if (last_point) {
-        const lp = grid.worldToScreen(last_point);
-        const unitVec = {
-          x: (x - lp.x) / pointDistance(lp, { x: x, y: y }),
-          y: (y - lp.y) / pointDistance(lp, { x: x, y: y }),
-        };
-        const perpUnitVec = { x: -unitVec.y, y: unitVec.x };
-        const midPoint = { x: (lp.x + x) * 0.5, y: (lp.y + y) * 0.5 };
-        const mid = { x: midPoint.x + perpUnitVec.x * 100, y: midPoint.y + perpUnitVec.y * 100 };
+        /* const lp = grid.toPoint(last_point);
+        const mp = grid.toPoint(mousePos); */
+        const mid = grid.toPoint({ x: (last_point.x + mousePos.x) * 0.5, y: (last_point.y + mousePos.y) * 0.5 });
         input.style.top = mid.y + "px";
         input.style.left = mid.x + "px";
+        /* const mp = grid.toPoint(mousePos);
+        input.style.top = mp.y + "px";
+        input.style.left = mp.x + "px"; */
         input.value = pointDistance(last_point, mousePos).toFixed(2);
         input.focus();
         input.select();
       }
-    } else if (currentTool === Tools.MOVE) {
-      if (isDragging) {
-        grid.offestX -= (x - dragStart.x) / grid.scaleX;
-        grid.offestY += (y - dragStart.y) / grid.scaleY;
-        dragStart = { x: x, y: y };
-      }
-      if (handleIsSelected) {
-        if (selectedHandleIndex instanceof Shape) {
-          selectedHandleIndex.calcularPropiedades();
-          const { XC: xc, YC: yc } = selectedHandleIndex.propiedades();
-          const dX = mousePos.x - xc;
-          const dY = mousePos.y - yc;
-          selectedHandleIndex.points.forEach((point) => {
-            point.x += dX;
-            point.y += dY;
-          });
-        } else {
-          selectedHandleIndex.x = mousePos.x;
-          selectedHandleIndex.y = mousePos.y;
-        }
+    } else if (currentTool === Tools.MOVE && handleIsSelected) {
+      if (selectedHandleIndex instanceof Shape) {
+        const dX = mousePos.x - prevMouse.x;
+        const dY = mousePos.y - prevMouse.y;
+        selectedHandleIndex.points.forEach((point) => {
+          point.x += dX;
+          point.y += dY;
+        });
+      } else {
+        selectedHandleIndex.x = mousePos.x;
+        selectedHandleIndex.y = mousePos.y;
       }
     }
-    redraw();
   };
+
   // Button
   document.getElementById("pencil").onclick = function () {
     switchTool(Tools.LINE);
-    canvas.style.cursor = "crosshair";
   };
   document.getElementById("arrows").onclick = function () {
     switchTool(Tools.MOVE);
-    canvas.style.cursor = "move";
   };
   document.getElementById("plus").onclick = function () {
     switchTool(Tools.ADD);
   };
   document.getElementById("scissors").onclick = function () {
     switchTool(Tools.CUT);
-    canvas.style.cursor = "crosshair";
+  };
+  document.getElementById("crosshairs").onclick = function () {
+    switchTool(Tools.ORIGIN);
+  };
+  document.getElementById("eye-slash").onclick = function () {
+    switchTool(Tools.VISIBILITY);
   };
   document.getElementById("anchor").onclick = function () {
     snap_enabled = !snap_enabled;
@@ -875,19 +888,37 @@ document.addEventListener("DOMContentLoaded", () => {
       };
     })(color_index);
   }
+
+  document.getElementById("clipboard").onclick = function () {
+    text.focus();
+    text.select();
+  };
+  document.getElementById("copy").onclick = function () {
+    var url = window.location.href.split("?")[0] + "?" + text.value;
+    window.open(url);
+  };
+  document.getElementById("refresh").onclick = function () {
+    if (confirm("Clear all polygon points and start over? This operation cannot be undone.")) {
+      switchTool(Tools.LINE);
+      shape = new Shape(false);
+      history.clear();
+      redraw();
+    }
+  };
+  document.getElementById("undo").onclick = function () {
+    undo();
+  };
+  document.getElementById("redo").onclick = function () {
+    redo();
+  };
+
   // Keyboard handler
   document.addEventListener("keydown", function (evt) {
     switch (evt.keyCode) {
       case 77: // "M"
       case 27: // <escape>
-        if (shape.points.length >= 2) {
-          shape.calcularPropiedades();
-          shapes.push(shape);
-        }
-        editor.removeChild(input);
-        shape = new Shape(true);
+        shape.last_point = null;
         switchTool(Tools.MOVE);
-        canvas.style.cursor = "move";
         break;
       case 76: // "L"
         switchTool(Tools.LINE);
@@ -915,158 +946,5 @@ document.addEventListener("DOMContentLoaded", () => {
         redraw();
         break;
     }
-  });
-
-  document.getElementById("zapatas2Form").addEventListener("submit", (event) => {
-    event.preventDefault();
-    const data = datosGenerales.getData();
-    markers = data.map((row) => {
-      return new Marker({ x: row.x, y: row.y }, row.column);
-    });
-    const cminx = markers.reduce((min, point) => (min < point.point.x ? min : point.point.x));
-    const cmaxx = markers.reduce((max, point) => (max > point.point.x ? max : point.point.x));
-
-    const cminy = markers.reduce((min, point) => (min < point.point.y ? min : point.point.y));
-    const cmaxy = markers.reduce((max, point) => (max > point.point.y ? max : point.point.y));
-
-    const dX = Math.abs(cminx - cmaxx);
-    const scaleX = canvas.width / dX;
-    grid.scaleX = scaleX;
-    grid.scaleY = scaleX;
-
-    const dY = Math.abs(cminy - cmaxy);
-    const scaleY = canvas.height / dY;
-    grid.scaleY = grid.scaleX = scaleX < scaleY ? scaleX * 0.9 : scaleY * 0.9;
-
-    grid.offestX = 0;
-    grid.offestY = 0;
-    const range = grid.screenToWorld({ x: canvas.width * 0.5, y: canvas.height * 0.5 });
-
-    grid.offestX = (cminx + cmaxx) * 0.5 - range.x;
-    grid.offestY = (cminy + cmaxy) * 0.5 - range.y;
-    redraw();
-  });
-
-  document.getElementById("calcularZapatas2").addEventListener("submit", (event) => {
-    event.preventDefault();
-    const octaveMatrix = (table, ...fields) => {
-      return (
-        "[" +
-        table
-          .getData()
-          .map((row) => {
-            return fields.map((field) => row[field]).join(" ");
-          })
-          .join(";") +
-        "]"
-      );
-    };
-    const formData = new FormData(event.target);
-    formData.append("column", octaveMatrix(datosGenerales, "column", "x", "y"));
-    formData.append("PD", octaveMatrix(datosGenerales, "column", "pd1", "pd2", "pd3"));
-    formData.append("PL", octaveMatrix(datosGenerales, "column", "pl1", "pl2", "pl3"));
-    formData.append("SISMO", octaveMatrix(datosGenerales, "column", "sismo1", "sismo2", "sismo3"));
-    formData.append(
-      "poligonos",
-      `struct(${shapes
-        .map((shape, index) => {
-          return `'poligono${index + 1}', [${[...shape.points, shape.points[0]].map((p) => `${p.x},${p.y}`).join(";")}]`;
-        })
-        .join(",")})`
-    );
-    const columns = datosGenerales.getData();
-    console.log(Object.fromEntries(formData));
-    fetch("/zapatas2", {
-      method: "POST",
-      body: formData,
-    })
-      .then(async (response) => {
-        const contentType = response.headers.get("Content-Type");
-        if (contentType && contentType.includes("application/octet-stream")) {
-          return response.arrayBuffer();
-        } else {
-          const error = await response.text();
-          return Promise.reject(error);
-        }
-      })
-      .then((matData) => {
-        const zapatas2 = mat4js.read(matData);
-        console.log(zapatas2);
-        Array.from(Array(6), (_, index) => index).forEach((_, index) => {
-          const traces = Object.values(zapatas2.data.resultados).map(({ XX, YY, ZZ, max, min }, poligonoN) => {
-            return {
-              x: XX, // X-axis data
-              y: YY, // Y-axis data
-              z: ZZ[index],
-              mode: "markers", // Scatter plot mode
-              marker: {
-                size: 2, // Size of the markers
-                color: ZZ[index], // Color of the markers, based on Z data
-                //colorscale: "Viridis", // Color scale
-                //colorscale: "Jet", // Color scale
-                /* showscale: true, // Show the color scale */
-                coloraxis: "coloraxis",
-              },
-              name: `<b>Poligono ${poligonoN + 1}<br>σ<sub>min</sub> = ${min[index].toFixed(3)}<br>σ<sub>max</sub> = ${max[index].toFixed(3)}</b><br>`,
-              hovertemplate: "<b>x</b>: %{x}<br>" + "<b>y</b>: %{y}<br>" + "<b>z</b>: %{marker.color:.4f}" + "<extra></extra>",
-              type: "scattergl", // 3D scatter plot type
-              /* type: "pointcloudgl", // 3D scatter plot type */
-            };
-          });
-          const markers = columns.map((row) => {
-            return {
-              x: [row.x],
-              y: [row.y],
-              text: [`${row.column}`],
-              mode: "markers+text",
-              type: "scattergl",
-              textposition: "top right",
-              textfont: {
-                family: "Arial",
-                size: 10,
-                color: "lightgrey",
-              },
-              showlegend: false,
-            };
-          });
-          const layout = {
-            xaxis: {
-              scaleanchor: "y",
-              scaleratio: 1,
-            },
-            yaxis: {
-              constrain: "domain",
-            },
-            coloraxis: {
-              colorscale: matlabColorScale, // Color scale
-              colorbar: {
-                title: {
-                  text: "Presion Admisible (Tn/m)",
-                  side: "right",
-                },
-              },
-            },
-            showlegend: true,
-            legend: { x: 0, y: 0.5 },
-            annotations: Object.values(zapatas2.data.resultados).map(({ XC: [xc], YC: [yc] }, index) => {
-              return {
-                x: xc,
-                y: yc,
-                xref: "x",
-                yref: "y",
-                text: `P${index + 1}`,
-                font: {
-                  color: "lightgrey",
-                  size: 10,
-                },
-                align: "left",
-                showarrow: false,
-              };
-            }),
-          };
-          // Plot the chart using Plotly
-          Plotly.react(`zapata${index + 1}`, [...traces, ...markers], layout, { responsive: false });
-        });
-      });
   });
 });
