+++ resources/views/matlab/admAnalisisEstructuralDeArmaduras.blade.php
@@ -1,25 +1,19 @@
 <x-app-layout>
     <x-slot name="header">
         <h2 class="font-semibold text-xl text-gray-800 dark:text-gray-200 leading-tight">
-            {{ __('Zapatas Grafico 2') }}
+            {{ __('Analisis Estructural de Armaduras') }}
         </h2>
     </x-slot>
     <style>
         .tabulator-cell:not(.tabulator-editable):not(.tabulator-calcs>.tabulator-cell) {
             background-color: #f2f2f2 !important;
         }
-
         #zapatas,
         #zapatas * {
             box-sizing: content-box;
-            color: 
         }
     </style>
     <link href="https://netdna.bootstrapcdn.com/font-awesome/4.6.3/css/font-awesome.css" rel="stylesheet">
-    <link href="https://unpkg.com/tabulator-tables/dist/css/tabulator.min.css" rel="stylesheet">
-    <script type="text/javascript" src="https://unpkg.com/tabulator-tables/dist/js/tabulator.min.js"></script>
-    <script src="https://unpkg.com/virtual-webgl@1.0.6/src/virtual-webgl.js"></script>
-    <script src="https://cdn.plot.ly/plotly-2.34.0.min.js" charset="utf-8"></script>
     <div class="py-12">
         <div class="container mx-auto w-full">
             <div class="flex flex-wrap">
@@ -28,28 +22,151 @@
                     <div class="bg-white dark:bg-gray-800 shadow-md rounded-lg p-6">
                         <h3 class="text-xl font-semibold text-gray-800 dark:text-gray-200 mb-4">Datos Generales</h3>
                         <div class="overflow-auto">
-                            <form id="zapatas2Form">
+                            <form id="zapatasForm">
                                 @csrf
                                 <table class="table-auto w-full text-gray-800 dark:text-white px-6">
-                                    <tr class="bg-white dark:bg-gray-800">
-                                        <td class="py-2 px-4" colspan="4">
-                                            <div id="datosGenerales"></div>
-                                        </td>
-                                    </tr>
-                                    <tr>
-                                        <th class="py-2 px-4 text-left" colspan="4">
-                                            <div class="input-group mb-2 text-left inline-block">
-                                                <button id="calcular"
-                                                    class="bg-blue-500 hover:bg-blue-400 text-white font-bold py-2 px-4 border-b-4 border-blue-700 hover:border-blue-500 rounded"
-                                                    type="submit">DISEÑAR</button>
-                                            </div>
-                                            <div class="input-group mb-2 text-left inline-block">
-                                                <button id="generarPDF"
-                                                    class="bg-blue-500 hover:bg-blue-400 text-white font-bold py-2 px-4 border-b-4 border-blue-700 hover:border-blue-500 rounded"
-                                                    type="button">PDF</button>
-                                            </div>
-                                        </th>
-                                    </tr>
+                                    <thead class="bg-white dark:bg-gray-800">
+                                        <tr class="text-center">
+                                            <th class="py-2 px-4">Nombre</th>
+                                            <th class="py-2 px-4">Simb.</th>
+                                            <th class="py-2 px-4">Entrada</th>
+                                            <th class="py-2 px-4">Unidad <br> Medida</th>
+                                        </tr>
+                                    </thead>
+                                    <tbody class="text-center">
+                                        <tr>
+                                            <td class="py-2 px-4">Modulo Elástico</td>
+                                            <td class="py-2 px-4">E</td>
+                                            <td class="py-2 px-4"><input type="number" name="modElastico"
+                                                    class="form-control w-full bg-gray-50 dark:bg-gray-700 text-gray-800 dark:text-gray-200 p-2 px-1 rounded-md"
+                                                    id="modElastico" min="0" value="210" required></td>
+                                            <td class="py-2 px-4">Kpa</td>
+                                        </tr>
+                                        <tr>
+                                            <td class="py-2 px-4">Área de la sección</td>
+                                            <td class="py-2 px-4">A</td>
+                                            <td class="py-2 px-4"><input type="number" name="aSeccion"
+                                                    class="form-control w-full bg-gray-50 dark:bg-gray-700 text-gray-800 dark:text-gray-200 p-2 px-1 rounded-md"
+                                                    id="aSeccion" min="0" value="210" required></td>
+                                            <td class="py-2 px-4">m<sup>2</sup></td>
+                                        </tr>
+                                        <tr>
+                                            <th class="text-xl py-2 px-4 text-left border-b border-gray-600"
+                                                colspan="4" scope="col">Nodos</th>
+                                        </tr>
+                                        <tr class="bg-white dark:bg-gray-800">
+                                            <th class="py-2 px-4">ID</th>
+                                            <th class="py-2 px-4"></th>
+                                            <th class="py-2 px-4">
+                                                <input type="number" name="nodoID"
+                                                    class="form-control w-full bg-gray-50 dark:bg-gray-700 text-gray-800 dark:text-gray-200 p-2 px-1 rounded-md"
+                                                    id="nodoID" min="0" required>
+                                            </th>
+                                            <th class="py-2 px-4"></th>
+                                        </tr>
+                                        <tr class="bg-white dark:bg-gray-800">
+                                            <th class="py-2 px-4">Posición</th>
+                                            <th class="py-2 px-4">x</th>
+                                            <th class="py-2 px-4">
+                                                <input type="number" name="posX"
+                                                    class="form-control w-full bg-gray-50 dark:bg-gray-700 text-gray-800 dark:text-gray-200 p-2 px-1 rounded-md"
+                                                    id="posX" placeholder="0" min="0" required>
+                                            </th>
+                                            <th class="py-2 px-4">m</th>
+                                        </tr>
+                                        <tr class="bg-white dark:bg-gray-800">
+                                            <th class="py-2 px-4">Posición</th>
+                                            <th class="py-2 px-4">y</th>
+                                            <th class="py-2 px-4">
+                                                <input type="number" name="posY"
+                                                    class="form-control w-full bg-gray-50 dark:bg-gray-700 text-gray-800 dark:text-gray-200 p-2 px-1 rounded-md"
+                                                    id="posY" placeholder="0" min="0" required>
+                                            </th>
+                                            <th class="py-2 px-4">m</th>
+                                        </tr>
+                                        <tr>
+                                            <th class="text-xl py-2 px-4 text-left border-b border-gray-600"
+                                                colspan="4" scope="col">Barras</th>
+                                        </tr>
+                                        <tr class="bg-white dark:bg-gray-800">
+                                            <th class="py-2 px-4">ID</th>
+                                            <th class="py-2 px-4"></th>
+                                            <th class="py-2 px-4">
+                                                <input type="number" name="barraID"
+                                                    class="form-control w-full bg-gray-50 dark:bg-gray-700 text-gray-800 dark:text-gray-200 p-2 px-1 rounded-md"
+                                                    id="barraID" min="0" required>
+                                            </th>
+                                            <th class="py-2 px-4"></th>
+                                        </tr>
+                                        <tr class="bg-white dark:bg-gray-800">
+                                            <th class="py-2 px-4">Nodo</th>
+                                            <th class="py-2 px-4">A</th>
+                                            <th class="py-2 px-4">
+                                                <input type="number" name="nodoAID"
+                                                    class="form-control w-full bg-gray-50 dark:bg-gray-700 text-gray-800 dark:text-gray-200 p-2 px-1 rounded-md"
+                                                    id="nodoAID" min="0"" required>
+                                            </th>
+                                            <th class="py-2 px-4">ID</th>
+                                        </tr>
+                                        <tr class="bg-white dark:bg-gray-800">
+                                            <th class="py-2 px-4">Nodo</th>
+                                            <th class="py-2 px-4">B</th>
+                                            <th class="py-2 px-4">
+                                                <input type="number" name="nodoBID"
+                                                    class="form-control w-full bg-gray-50 dark:bg-gray-700 text-gray-800 dark:text-gray-200 p-2 px-1 rounded-md"
+                                                    id="nodoBID" min="0"" required>
+                                            </th>
+                                            <th class="py-2 px-4">ID</th>
+                                        </tr>
+                                        <tr>
+                                            <th class="text-xl py-2 px-4 text-left border-b border-gray-600"
+                                                colspan="4" scope="col">Soportes</th>
+                                        </tr>
+                                        <tr>
+                                            <th class="text-xl py-2 px-4 text-left border-b border-gray-600"
+                                                colspan="4" scope="col">Fuerzas</th>
+                                        </tr>
+                                        <tr class="bg-white dark:bg-gray-800">
+                                            <th class="py-2 px-4">ID</th>
+                                            <th class="py-2 px-4"></th>
+                                            <th class="py-2 px-4">
+                                                <input type="number" name="nodoID"
+                                                    class="form-control w-full bg-gray-50 dark:bg-gray-700 text-gray-800 dark:text-gray-200 p-2 px-1 rounded-md"
+                                                    id="nodoID" min="0" required>
+                                            </th>
+                                            <th class="py-2 px-4"></th>
+                                        </tr>
+                                        <tr class="bg-white dark:bg-gray-800">
+                                            <th class="py-2 px-4">Magnitud</th>
+                                            <th class="py-2 px-4">X</th>
+                                            <th class="py-2 px-4">
+                                                <input type="number" name="magnitudX"
+                                                    class="form-control w-full bg-gray-50 dark:bg-gray-700 text-gray-800 dark:text-gray-200 p-2 px-1 rounded-md"
+                                                    id="magnitudX" min="0" required>
+                                            </th>
+                                            <th class="py-2 px-4">kN</th>
+                                        </tr>
+                                        <tr class="bg-white dark:bg-gray-800">
+                                            <th class="py-2 px-4">Magnitud</th>
+                                            <th class="py-2 px-4">Y</th>
+                                            <th class="py-2 px-4">
+                                                <input type="number" name="magnitudY"
+                                                    class="form-control w-full bg-gray-50 dark:bg-gray-700 text-gray-800 dark:text-gray-200 p-2 px-1 rounded-md"
+                                                    id="magnitudY" min="0" required>
+                                            </th>
+                                            <th class="py-2 px-4">kN</th>
+                                        </tr>
+                                        <!-- Agregar más filas según sea necesario -->
+                                        <tr>
+                                            <th class="py-2 px-4">
+                                                <div class="input-group mb-2 text-left">
+                                                    <button id="calcular"
+                                                        class="bg-blue-500 hover:bg-blue-400 text-white font-bold py-2 px-4 border-b-4 border-blue-700 hover:border-blue-500 rounded"
+                                                        type="submit">CARGAR</button>
+                                                </div>
+                                            </th>
+                                        </tr>
+                                    </tbody>
                                 </table>
                             </form>
                         </div>
@@ -60,12 +177,6 @@
                 <div class="w-full md:w-2/3 px-4 mt-4 md:mt-0">
                     <div class="bg-white dark:bg-gray-800 shadow-md rounded-lg p-6">
                         <h3 class="text-xl font-semibold text-gray-800 dark:text-gray-200 mb-4">Resultados</h3>
-                        <form id="calcularZapatas2">
-                            @csrf
-                            <button
-                                class="bg-blue-500 hover:bg-blue-400 text-white font-bold py-2 px-4 border-b-4 border-blue-700 hover:border-blue-500 rounded"
-                                type="submit">CALCULAR</button>
-                        </form>
                         <div class="overflow-x-auto" id="resultados">
                             <table class="min-w-full text-gray-800 dark:text-white">
                                 <thead>
@@ -80,7 +191,7 @@
                                             <div>
                                                 <form>
                                                     <textarea class="hidden" name="coords" readonly></textarea>
-                                                    <input class="hidden" type="range" name="zoom" min="5"
+                                                    <input type="range" name="zoom" min="5"
                                                         max="40">
                                                 </form>
                                                 <ul class="flex gap-4">
@@ -96,16 +207,12 @@
                                                     <li id="anchor"><i class="fa fa-anchor"
                                                             title="S: Toggle Grid Snap"></i>
                                                     </li>
-                                                    <li><input type="number" name="snap"
-                                                            class="form-control w-3/4 bg-gray-50 dark:bg-gray-700 text-gray-800 dark:text-gray-200 p-2 px-1 rounded-md"
-                                                            id="snap" value="1" step="any" min="0"
-                                                            required></i>
-                                                    </li>
                                                 </ul>
                                                 <ul class="flex gap-4 hidden">
                                                     <li id="undo"><i class="fa fa-undo" title="U: Undo"></i>
                                                     <li id="redo"><i class="fa fa-repeat" title="R: Redo"></i>
-                                                    <li id="refresh"><i class="fa fa-trash-o" title="Delete All"></i>
+                                                    <li id="refresh"><i class="fa fa-trash-o"
+                                                            title="Delete All"></i>
                                                     </li>
                                                 </ul>
                                                 {{-- <ol class="flex gap-4">
@@ -162,38 +269,7 @@
                                     </tr>
                                 </thead>
                                 <tbody id="polygons">
-                                </tbody>
-                                <tbody>
-                                    <tr class="bg-gray-100 dark:bg-gray-600">
-                                        <td class="py-2 px-4" colspan="4">
-                                            <div id="zapata1"></div>
-                                        </td>
-                                    </tr>
-                                    <tr class="bg-gray-100 dark:bg-gray-600">
-                                        <td class="py-2 px-4" colspan="4">
-                                            <div id="zapata2"></div>
-                                        </td>
-                                    </tr>
-                                    <tr class="bg-gray-100 dark:bg-gray-600">
-                                        <td class="py-2 px-4" colspan="4">
-                                            <div id="zapata3"></div>
-                                        </td>
-                                    </tr>
-                                    <tr class="bg-gray-100 dark:bg-gray-600">
-                                        <td class="py-2 px-4" colspan="4">
-                                            <div id="zapata4"></div>
-                                        </td>
-                                    </tr>
-                                    <tr class="bg-gray-100 dark:bg-gray-600">
-                                        <td class="py-2 px-4" colspan="4">
-                                            <div id="zapata5"></div>
-                                        </td>
-                                    </tr>
-                                    <tr class="bg-gray-100 dark:bg-gray-600">
-                                        <td class="py-2 px-4" colspan="4">
-                                            <div id="zapata6"></div>
-                                        </td>
-                                    </tr>
+
                                 </tbody>
                             </table>
                         </div>
